Develop

+++ app/src/main/kotlin/site/paulo/localchat/data/model/chatgeo/User.kt
@@ -3,7 +3,7 @@ package site.paulo.localchat.data.model.chatgeo
 data class User(
     val name: String = "",
     val chats: Map<String, Boolean> = emptyMap(),
-    val age: Int = 0,
+    val age: Long = 0L,
     val email: String = "",
     val gender: String = "",
     val profilePic: String = "")
